Updated file again
I'm sorry that I change something here again, but in German there is the politeness (you, your) and the form of direct address and since this is a new project or rather a start-up and it is more likely to appeal to new, young customers, I chose the direct form to highlight the Instatus image. In addition, I made small unformulations in order to sound more serious.